Refactor to allow easier import
