+++ clients/admin-ui/src/features/system/SystemFormTabs.tsx
@@ -7,9 +7,9 @@ import { useAppDispatch, useAppSelector } from "~/app/hooks";
 import DataTabs, { type TabData } from "~/features/common/DataTabs";
 import { useInterzoneNav } from "~/features/common/hooks/useInterzoneNav";
 import { DEFAULT_TOAST_PARAMS } from "~/features/common/toast";
+import PrivacyDeclarationStep from "~/features/system/privacy-declarations/PrivacyDeclarationStep";
 import { System } from "~/types/api";
 
-import PrivacyDeclarationStep from "./PrivacyDeclarationStep";
 import { selectActiveSystem, setActiveSystem } from "./system.slice";
 import SystemInformationForm from "./SystemInformationForm";
 import UnmountWarning from "./UnmountWarning";
