+++ clients/admin-ui/src/features/system/privacy-declarations/PrivacyDeclarationManager.tsx
@@ -1,25 +1,10 @@
-import {
-  Box,
-  Button,
-  Heading,
-  Stack,
-  Text,
-  Tooltip,
-  useToast,
-} from "@fidesui/react";
-import { SerializedError } from "@reduxjs/toolkit";
-import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query/fetchBaseQuery";
-import NextLink from "next/link";
+import { Box, Button, Stack, Tooltip, useToast } from "@fidesui/react";
 import { useMemo, useState } from "react";
 
-import { useAppDispatch } from "~/app/hooks";
-import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
 import { PrivacyDeclaration, System } from "~/types/api";
 
-import { errorToastParams, successToastParams } from "../common/toast";
 import PrivacyDeclarationAccordion from "./PrivacyDeclarationAccordion";
-import ConnectedPrivacyDeclarationForm from "./PrivacyDeclarationForm";
-import { setActiveSystem, useUpdateSystemMutation } from "./system.slice";
+import { DataProps, PrivacyDeclarationForm } from "./PrivacyDeclarationForm";
 
 type FormValues = PrivacyDeclaration;
 
@@ -33,13 +18,17 @@ const transformFormValuesToDeclaration = (
 
 interface Props {
   system: System;
+  onCollision: () => void;
+  onSave: (privacyDeclarations: PrivacyDeclaration[]) => Promise<boolean>;
 }
 
-const PrivacyDeclarationStep = ({ system }: Props) => {
+const PrivacyDeclarationManager = ({
+  system,
+  onCollision,
+  onSave,
+  ...dataProps
+}: Props & DataProps) => {
   const toast = useToast();
-  const dispatch = useAppDispatch();
-  const [updateSystemMutationTrigger, { isLoading }] =
-    useUpdateSystemMutation();
 
   const [showNewForm, setShowNewForm] = useState(false);
   const [newDeclaration, setNewDeclaration] = useState<
@@ -60,50 +49,18 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
       accordionDeclarations.filter((d) => d.data_use === values.data_use)
         .length > 0
     ) {
-      toast(
-        errorToastParams(
-          "A declaration already exists with that data use in this system. Please supply a different data use."
-        )
-      );
+      onCollision();
       return true;
     }
     return false;
   };
 
-  const save = async (updatedDeclarations: PrivacyDeclaration[]) => {
+  const handleSave = async (updatedDeclarations: PrivacyDeclaration[]) => {
     const transformedDeclarations = updatedDeclarations.map((d) =>
       transformFormValuesToDeclaration(d)
     );
-    const systemBodyWithDeclaration = {
-      ...system,
-      privacy_declarations: transformedDeclarations,
-    };
-
-    const handleResult = (
-      result:
-        | { data: System }
-        | { error: FetchBaseQueryError | SerializedError }
-    ) => {
-      if (isErrorResult(result)) {
-        const errorMsg = getErrorMessage(
-          result.error,
-          "An unexpected error occurred while updating the system. Please try again."
-        );
-
-        toast(errorToastParams(errorMsg));
-        return false;
-      }
-      toast.closeAll();
-      toast(successToastParams("Data use case saved"));
-      dispatch(setActiveSystem(result.data));
-      return true;
-    };
-
-    const updateSystemResult = await updateSystemMutationTrigger(
-      systemBodyWithDeclaration
-    );
-
-    return handleResult(updateSystemResult);
+    const success = await onSave(transformedDeclarations);
+    return success;
   };
 
   const handleEditDeclaration = async (
@@ -122,7 +79,7 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
     const updatedDeclarations = accordionDeclarations.map((dec) =>
       dec.data_use === oldDeclaration.data_use ? updatedDeclaration : dec
     );
-    const success = await save(updatedDeclarations);
+    const success = await handleSave(updatedDeclarations);
     return success;
   };
 
@@ -134,7 +91,7 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
     toast.closeAll();
     setNewDeclaration(values);
     const updatedDeclarations = [...accordionDeclarations, values];
-    const success = await save(updatedDeclarations);
+    const success = await handleSave(updatedDeclarations);
     return success;
   };
 
@@ -148,32 +105,18 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
     (system.privacy_declarations.length === 0 && !showNewForm);
 
   return (
-    <Stack spacing={3} data-testid="privacy-declaration-step">
-      <Heading as="h3" size="md">
-        Data uses
-      </Heading>
-      <Text fontSize="sm">
-        Data Uses describe the business purpose for which the personal data is
-        processed or collected. Within a Data Use, you assign which categories
-        of personal information are collected for this purpose and for which
-        categories of data subjects. To update the available categories and
-        uses, please visit{" "}
-        <NextLink href="/taxonomy" passHref>
-          <Text as="a" color="complimentary.600">
-            Manage taxonomy
-          </Text>
-        </NextLink>
-        .
-      </Text>
+    <Stack spacing={3}>
       <PrivacyDeclarationAccordion
         privacyDeclarations={accordionDeclarations}
         onEdit={handleEditDeclaration}
+        {...dataProps}
       />
       {showNewForm ? (
         <Box backgroundColor="gray.50" p={6} data-testid="new-declaration-form">
-          <ConnectedPrivacyDeclarationForm
+          <PrivacyDeclarationForm
             initialValues={newDeclaration}
             onSubmit={saveNewDeclaration}
+            {...dataProps}
           />
         </Box>
       ) : null}
@@ -188,7 +131,6 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
             <Button
               colorScheme="primary"
               size="xs"
-              isLoading={isLoading}
               data-testid="add-btn"
               onClick={handleShowNewForm}
               disabled={showNewForm && !newDeclaration}
@@ -202,4 +144,4 @@ const PrivacyDeclarationStep = ({ system }: Props) => {
   );
 };
 
-export default PrivacyDeclarationStep;
+export default PrivacyDeclarationManager;
